requirements.txt
crewai
openai
ddgs
python-dotenv
langchain-openai 